LTP - report a pass if a corefile is found
The core files may be dumped because of a problem in kernel, but also
because of a completely unrelated one. Because of the later, we don't
want to report a warning to the developer if a core file is found but
all the tests passed.

Changing only the reported result but keeping the core files in the job
allows for debugging / reporting a problem to unrelated component if
it's needed.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>