Fix kernel install on distros using BLS
The previously added new-kernel-pkg doesn't work on these systems.
Detect the missing script and use the recommended way to install the
kernel there.

Use grubby to set the default boot entry because there is an issue with
either grub2-set-default or kernel-install which makes us unable to set
a custom kernel as the default using grub2-set-default. The workaround
is noted in the comment and can be removed after the problem is
resolved.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>