kpkginstall: adhere to test standards
skt expects the task to abort in case the issue is not a real test
failure. If the caller eg. forgets to add the right parameters, it falls
under this category.

Change the "FAIL" reporting to aborting the recipe. Also adjust the name
of the "install" subtask to make sure it doesn't get confused with the
test installation subtask which some tests have (logs from the test
installation aren't added to the reports).

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>