LTP: adhere to test standards
We expect the task to abort the recipe in case there's an issue that is
not a test failure; to be able to check and fix the problems and not
report them to developers since it's not their fault. "We had issues
patching the testsuite" definitely falls in this category.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>