kpkginstall: Stop requiring KVER parameter
Instead of requiring the KVER parameter being passed through the
environment in distribution/kpkginstall, extract the kernel version from
the kernel image filename in the tarball.

This simplifies the interface and workings of kpet, so that it only has
to deal with a tarball, and doesn't have to know which kernel version
it's dealing with.