kpkginstall: abort when the issue is caused by the OS
Inability to extract the tarball would most likely indicate lack of disk
space or denied permissions for the user, not a problem with the kernel.
Similarly, having a different kernel booted after the installation
completed could be (and we did run into this situation) caused by a bug
in grub.

We'll be transitioning to treating this task as a normal test. Therefore
the only thing that should report a real failure is a kernel panic
during the boot.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>