Add explicit return to rpminstall function
Implicit return gives us the code of lastly executed command, which in
this case is the kernel-firmware installation we don't really care
about. So we would abort the task with a "failed to install kernel"
message despite the fact that the kernel was successfully installed and
we should continue execution.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>