Add installation of linux-firmware
The kernel-firmware was renamed to linux-firmware in newer releases,
which is why we were unable to install the package. We should check
which package is available and always install it.

Also, yum/dnf is has this fancy "feature" that it returns 0 if we search
for nonexistent package, so we have to work around it with grep.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>