Only list available packages
Showduplicates is nice to get us older packages than currently installed
kernel. However this means that after the installation, our kernel is
listed both as "installed" and "available" and thus shows up twice in
the version extraction, leading into having a version like
"4.20.8-100.fc28\n4.20.8-100.fc28"; which is not what we want.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>