Fixup package name grepping
We are runnning the task in Beaker which means Beaker repository with
tests is added by default. There are tests for kernel-firmware too which
means the grep always succeeds even though there is no package with that
exact name. Modify the grep to only check the match from the start of
the line and match the full package name to avoid false matches.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>