Fix kpkginstall for kernel-rt
The package is not called "kernel" so all the "yum install kernel", "yum
list kernel" etc. fail. Try to extract a kernel-like package from the
repo and use this package name instead of hardcoded "kernel".

Right now the only different package name we know about is the mentioned
kernel-rt. If there's more of them, we need to add them into the "list"
we check against.

The other option would be to pass the package name as another parameter
to the task. I opted against that for now because we worked hard to get
rid of all parameters we don't need.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>