ltp/include: disable distro exclusion in upstream kernel testing
As Veronika pointed that the distro exclusion cann't work properly in
upstream kernel testing on RHEL. Becasue sometimes a RHEL kernel bug
maybe has already been fixed in upstream, the distro exclusion still
trying to kick-out it as a known issue.

To solve that we could simply try to detect whether a distro installs the
upstream kernel, if yes, we just give up the filtering rule at this case.
It will be eventually separate the upstream and downstream. Problem could
be located in any part of these: [UPSTREAM] [RHEL8] [RHEL7] [RHEL6] [RHEL5]
Btw, the [UPSTREAM] part includes: CKI, ARK and Mainline kernel.

Note:
  osver_in_range is just as workaround for unfix knownissue filtering, once
  an unfix problem is being fixed there, we should move it to fixed issue
  part and change the rule from distro exclusion to kernel NVR comparing.

So, we'd better follow these principles to exlucde testcase in future:
  1. Upstream kernel bug use its kernel-nvr ranges
  2. RHEL kernel bug(fixed) use its kernel-nvr ranges
  3. RHEL kernel bug(unfix) use distro exclustion first, then move to
     kernel-nvr ranges once it has been fixed.
  4. Userspace package bug use itself package-nvr ranges

Signed-off-by: Li Wang <liwang@redhat.com>
Acked-by: Jan Stancek <jstancek@redhat.com>