warn/abort recipe on infrastructure issues
This implements FASTMOVING-591.

When a kernel installation fails, the whole recipe isn't aborted. This
is not the behavior we want. We need warn/abort returned and the job to
be resubmitted.
This patch adds some explicit recipe aborts whenever kernel
installation fails; they seem to be missing by accident.
We also add report-result calls.

This patch will need a followup: we don't really return any special
return codes when we have a failure that's likely an issue of this task
(e.g. script fails to extract kversion from string).

V2
* report warn so skt knows what to do
V3:
* "exit 1" would report fail instead of warn, so report warn, abort recipe,
  then do "exit 0"

Signed-off-by: Jakub Racek <jracek@redhat.com>