perf: adding a new testsuite (wrapper)
The perftool-testsuite should become a part of CKI perf testing as well.
It provides a different approach of perf testing than the internal-testsuite.
The wrapper is based on the wrapper used at Red Hat, but adjusted for the
CKI usage. It contains a blacklist to mask out known issues and unstable
tests, since CI should be stable.