perf/perftool-testsuite: adoption of the upstream new runmode model (#299)
The perftool-testsuite now supports three runmodes which differ in which
testcases are run and how deep the testsuite does the testing.

BASIC mode is designed for CI. STANDARD mode for normal testing and
EXPERIMENTAL mode is for playing with unstable tests. This should prevent
CKI from failing due to new unstable tests added to the testsuite, like
it happened recently with perf-script tests.