Debug kernel install support for kpkginstall (#335)
* Store PACKAGE_NAME across reboots

Signed-off-by: Major Hayden <major@redhat.com>

* Make temp directory to hold configs

Signed-off-by: Major Hayden <major@redhat.com>

* Store KVER across reboots

Signed-off-by: Major Hayden <major@redhat.com>

* Refactor kernel version check after reboot

Set up an array of acceptable version strings and check the uname output
against that array. Also, use less nesting of if/else statements by
ending immediately if a call trace is found.

Also, add emojis. Because why not? 🤷

Signed-off-by: Major Hayden <major@redhat.com>

* Add debug kernel install support

Ensure that debug kernels can be properly installed and selected as the
default after reboot. Also, add a new list of kernel version strings to
use when checking for a debug kernel after reboot.

Signed-off-by: Major Hayden <major@redhat.com>