Don't depend on the "service" command in tests
Switch calls to the "service" command to the BeakerLib equivalent
function. Because there's no rlServiceRestart, all of the "service xxx
restart" have been changed to two calls: "rlServiceStop xxx" and
"rlServiceStart xxx". Hopefully this doesn't change anything or break
any tests, but it's good to be aware that there could be a difference.

NB: The following files can be ignored because they are already using
systemctl when available, and only using service for older RHEL
versions:

* kdump/include/runtest.sh
* networking/route/pmtu/common/network.sh
* networking/firewall/000infralib/src/lib/service.sh
* networking/common/network.sh
* networking/common/service.sh
* networking/bridge/common/lib.sh