Use vmlinuz-KVR when using grubby
When using `grubby --set-default`, the argument must be in the
form of `vmlinuz-KVR` instead of just `KVR`. Without this change,
the kernel won't show up as the primary kernel at boot time. The
originally installed kernel will boot instead.

Also, remove the `/dev/null` redirect so we can see errors from
`grubby` and `zipl` in the future.

Fixes FASTMOVING-1315.

Signed-off-by: Major Hayden <major@redhat.com>