fs/xfs/xfstests: whitelist dmesg string MAX_LOCKDEP_CHAIN_HLOCKS
After we swtich to run tests on debug kernel, it's easy to hit this
warning during dmesg check of beaker by random subcase of xfstests.
xfstests upstream has whitelist this string as it is a false alarm,
not a real issue.

We can't skip subcase which is hitting this warning because any
stressful subcase could hit this. So bring generic/374 back, which
was wrongly skipped when we first hit this warning.

This solution suggested by Rachel should work according to:
https://git.beaker-project.org/cgit/restraint/tree/plugins/report_result.d/01_dmesg_check#n29

Signed-off-by: xiangcai <jencce2002@gmail.com>