firmware: set fwts version to V20.03.00
Up until now fwts has been running with the master branch.
This has been causing failures since some of the changes in the patches
no longer apply cleanly to the current master.
Let's switch to a fixed tag instead so the test is more stable and not
so sensible to upstream changes.

Signed-off-by: Erico Nunes <ernunes@redhat.com>