ftrace-tracer: Limit the output of function and function_graph
In busy systems function and function_graph could have a lot of output
and when we cat the trace data the ring buffer increases faster than
cat, which results in infinite waiting. So now switch all to show only
the top lines to reduce the chance of aborts in test.

Signed-off-by: Ziqian SUN (Zamir) <zsun@redhat.com>