Fix detection of expected kernel NVR
We grab the expected NVR, then overwrite it with the NVR of currently
running kernel, and then compare the NVR of the currently running kernel
with the NVR of the currently running kernel. Very useful. This leads to
missing bugs when the wrong kernel was booted, such as the s390x default
kernel detection.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>